+++ internal/database/model/business_card.go
@@ -1,22 +1,6 @@
 package model
 
-import (
-	"time"
-)
-
-// UserはユーザーテーブルのORMモデルです。
-type User struct {
-	ID                              string                           `gorm:"primaryKey"`
-	BusinessCards                   []BusinessCard                   `gorm:"foreignKey:User"`
-	PersonalBusinessCardBackgrounds []PersonalBusinessCardBackground `gorm:"foreignKey:User"`
-	PersonalThreeDimentionalModels  []PersonalThreeDimentionalModel
-	SpeakingAssets                  []SpeakingAsset
-	RecordedVoicePath               string
-	RecordedModelPath               string
-	CreatedAt                       time.Time
-	DeletedAt                       time.Time
-	IsToured                        bool
-}
+import "time"
 
 // BusinessCardは名刺テーブルのORMモデルです。
 type BusinessCard struct {
@@ -90,43 +74,4 @@ type PersonalBusinessCardBackground struct {
 	User                    string
 	ColorCode               string
 	ImagePath               string
-}
-
-// ThreeDimentionalModelは3DモデルテーブルのORMモデルです。
-type ThreeDimentionalModel struct {
-	ID       string `gorm:"primaryKey"`
-	ARAssets []ARAsset
-}
-
-// ThreeDimentionalModelTemplateは3DモデルテンプレートテーブルのORMモデルです。
-// ThreeDimentionalModelの排他的サブタイプです。
-type ThreeDimentionalModelTemplate struct {
-	ID   string `gorm:"primaryKey"`
-	Path string
-}
-
-// PersonalThreeDimentionalModelはユーザー定義3DモデルテーブルのORMモデルです。
-// ThreeDimentionalModelの排他的サブタイプです。
-type PersonalThreeDimentionalModel struct {
-	ID     string `gorm:"primaryKey"`
-	UserID string
-	Path   string
-}
-
-// SpeakingAssetは音声アセットテーブルのORMモデルです。
-type SpeakingAsset struct {
-	ID          string `gorm:"primaryKey"`
-	ARAssets    []ARAsset
-	UserID      string
-	Description string
-	AudioPath   string
-}
-
-// ARAssetはARアセットテーブルのORMモデルです。
-type ARAsset struct {
-	ID                      string         `gorm:"primaryKey"`
-	BusinessCards           []BusinessCard `gorm:"foreignKey:ARAsset"`
-	UserID                  string
-	SpeakingAssetID         string
-	ThreeDimentionalModelID string
 }
